MMAKE movingFurniture to fewer re-rendered
